Encabezado de Clases y union de la clase para manejo de archivos
