+++ nbproject/private/private.xml
@@ -3,10 +3,8 @@
     <editor-bookmarks xmlns="http://www.netbeans.org/ns/editor-bookmarks/2" lastBookmarkId="0"/>
     <open-files xmlns="http://www.netbeans.org/ns/projectui-open-files/2">
         <group>
-            <file>file:/home/jesfrin/NetBeansProjects/LaberintoTeoria/src/gnz/backend/Matriz/ManejadorMatriz.java</file>
-            <file>file:/home/jesfrin/NetBeansProjects/LaberintoTeoria/src/gnz/backend/Matriz/Celda.java</file>
-            <file>file:/home/jesfrin/NetBeansProjects/LaberintoTeoria/src/gnz/backend/archivos/ManejadorDeArchivos.java</file>
-            <file>file:/home/jesfrin/NetBeansProjects/LaberintoTeoria/src/gnz/frontend/ManejadorDePanel.java</file>
+            <file>file:/home/bryan/NetBeansProjects/OM/OM_V1.0/src/gnz/backend/objetoMovil/Archivos.java</file>
+            <file>file:/home/bryan/NetBeansProjects/OM/OM_V1.0/src/gnz/backend/Matriz/Run.java</file>
         </group>
     </open-files>
 </project-private>
