+++ src/gnz/backend/Matriz/Celda.java
@@ -1,7 +1,5 @@
 /*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
+ * Clase Celda, que es usada por la Matriz en el Panel Visual
  */
 package gnz.backend.Matriz;
