+++ src/gnz/backend/archivos/ManejadorDeArchivos.java
@@ -1,17 +1,17 @@
 /*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
+ * Clase encargada del manejo de archivos
  */
 package gnz.backend.archivos;
 
 import gnz.backend.Matriz.ManejadorMatriz;
 import gnz.frontend.FrameOM;
 import java.awt.Color;
+import java.io.BufferedReader;
 import java.io.File;
 import java.io.FileInputStream;
 import java.io.FileNotFoundException;
 import java.io.FileOutputStream;
+import java.io.FileReader;
 import java.io.ObjectInputStream;
 import java.io.ObjectOutputStream;
 import java.util.logging.Level;
@@ -19,6 +19,7 @@ import java.util.logging.Logger;
 import javax.swing.JFileChooser;
 import javax.swing.JOptionPane;
 import javax.swing.JPanel;
+import javax.swing.filechooser.FileNameExtensionFilter;
 
 /**
  *
@@ -114,6 +115,33 @@ public class ManejadorDeArchivos {
 
     public void setMapaEstaGuardado(boolean mapaEstaGuardado) {
         this.mapaEstaGuardado = mapaEstaGuardado;
+    }
+    
+    
+    /*Lectura y escritura basica de archivos de texto*/
+    
+    public String leer(String cadena){
+        String texto="";
+        try {
+            BufferedReader bf= new BufferedReader(new FileReader(cadena));
+            String temp="";
+            String bfReader;
+            while ((bfReader=bf.readLine())!=null) {
+                temp=temp+bfReader+"\n";
+            }
+            texto=temp;
+        } catch (Exception e) {
+        }
+        return texto;
+    }
+    
+    public File Abrir(){
+        JFileChooser chooser=new JFileChooser();
+        FileNameExtensionFilter filter = new FileNameExtensionFilter("obmo", "OBMO");
+        chooser.setFileFilter(filter);
+        chooser.showOpenDialog(null);
+        File f=chooser.getSelectedFile();
+        return f;
     }
 
 }
