+++ src/gnz/backend/objetoMovil/Movimiento.java
@@ -1,7 +1,5 @@
 /*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
+ * Clase para atribuir los movimientos base al OM
  */
 package gnz.backend.objetoMovil;
